Persistence Intervals and Diagrams

PersistenceDiagramsBase.PersistenceInterval — Type

PersistenceInterval

The type that represents a persistence interval. It behaves exactly like a Tuple{Float64, Float64}, but can have meta data attached to it. The metadata is accessible with getproperty.

Example

julia> interval = PersistenceInterval(1, Inf; meta1=:a, meta2=:b)
[1.0, ∞) with:
 meta1: Symbol
 meta2: Symbol

julia> birth(interval), death(interval), persistence(interval)
(1.0, Inf, Inf)

julia> isfinite(interval)
false

julia> propertynames(interval)
(:birth, :death, :meta1, :meta2)

julia> interval.meta1
:a

PersistenceDiagramsBase.birth — Function

birth(interval)

Get the birth time of interval.

PersistenceDiagramsBase.death — Function

death(interval)

Get the death time of interval.

PersistenceDiagramsBase.persistence — Function

persistence(interval)

Get the persistence of interval, which is equal to death - birth.

PersistenceDiagramsBase.representative — Function

representative(interval::PersistenceInterval)

Get the representative (co)cycle attached to interval, if it has one.

PersistenceDiagramsBase.birth_simplex — Function

birth_simplex(interval::PersistenceInterval)

Get the critical birth simplex of interval, if it has one.

PersistenceDiagramsBase.death_simplex — Function

death_simplex(interval::PersistenceInterval)

Get the critical death simplex of interval, if it has one.

Note

An infinite interval's death simplex is nothing.

PersistenceDiagramsBase.PersistenceDiagram — Type

PersistenceDiagram{P<:PersistenceInterval} <: AbstractVector{P}

Type for representing persistence diagrams. Behaves exactly like a vector of PersistenceIntervals, but is can have metadata attached to it and supports pretty printing and plotting.

Can be used as a table with any function that uses the Tables.jl interface.

Example

julia> diagram = PersistenceDiagram([(1, 3), (3, 4), (1, Inf)]; dim=1, custom_metadata=:a)
3-element 1-dimensional PersistenceDiagram:
 [1.0, 3.0)
 [3.0, 4.0)
 [1.0, ∞)

julia> diagram[1]
[1.0, 3.0)

julia> sort(diagram; by=persistence, rev=true)
3-element 1-dimensional PersistenceDiagram:
 [1.0, ∞)
 [1.0, 3.0)
 [3.0, 4.0)

julia> propertynames(diagram)
(:dim, :custom_metadata)

julia> dim(diagram)
1

julia> diagram.custom_metadata
:a

Missing docstring.

Missing docstring for PersistenceDiagrams.table. Check Documenter's build log for details.

PersistenceDiagrams.barcode — Method

barcode(diagram)

Plot the barcode plot of persistence diagram or multiple diagrams in a collection. The infinity keyword argument determines where the infinity line is placed. If unset, the function tries to use threshold(diagram), or guess a good position to place the line at.

Distances Between Persistence Diagrams

PersistenceDiagrams.Bottleneck — Type

Bottleneck

Use this object to find the bottleneck distance or matching between persistence diagrams. The distance value is equal to

\[W_\infty(X, Y) = \inf_{\eta:X\rightarrow Y} \sup_{x\in X} ||x-\eta(x)||_\infty,\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the bottleneck distance requires $\mathcal{O}(n^2)$ space. Be careful when computing distances between very large diagrams!

Usage

Bottleneck()(left, right[; matching=false]): find the bottleneck matching (if matching=true) or distance (if matching=false) between persistence diagrams left and right

Example

julia> left = PersistenceDiagram([(1.0, 2.0), (5.0, 8.0)]);

julia> right = PersistenceDiagram([(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)]);

julia> Bottleneck()(left, right)
2.0

julia> Bottleneck()(left, right; matching=true)
Bottleneck Matching with weight 2.0:
 [5.0, 8.0) => [5.0, 10.0)

PersistenceDiagrams.Wasserstein — Type

Wasserstein(q=1)

Use this object to find the Wasserstein distance or matching between persistence diagrams. The distance value is equal to

\[W_q(X,Y)=\left[\inf_{\eta:X\rightarrow Y}\sum_{x\in X}||x-\eta(x)||_\infty^q\right],\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the Wasserstein distance requires $\mathcal{O}(n^2)$ space. Be careful when computing distances between very large diagrams!

Usage

Wasserstein(q=1)(left, right[; matching=false]): find the Wasserstein matching (if matching=true) or distance (if matching=false) between persistence diagrams left and right.

Example

julia> left = PersistenceDiagram([(1.0, 2.0), (5.0, 8.0)]);

julia> right = PersistenceDiagram([(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)]);

julia> Wasserstein()(left, right)
3.0

julia> Wasserstein()(left, right; matching=true)
Matching with weight 3.0:
 [1.0, 2.0) => [1.0, 2.0)
 [3.0, 3.0) => [3.0, 4.0)
 [5.0, 8.0) => [5.0, 10.0)

PersistenceDiagrams.matching — Function

matching(::MatchingDistance, left, right)
matching(::Matching)

Get the matching between persistence diagrams left and right.

See also

weight

Bottleneck

Wasserstein

PersistenceDiagrams.weight — Function

weight(::MatchingDistance, left, right)
weight(::Matching)

Get the weight of the matching between persistence diagrams left and right.

See also

matching

Bottleneck

Wasserstein

Persistence Diagram Vectorization

PersistenceDiagrams.PersistenceImage — Type

PersistenceImage

PersistenceImage provides a vectorization method for persistence diagrams. Each point in the diagram is first transformed into birth, persistence coordinates. Then, it is weighted by a weighting function and widened by a distribution (default: gaussian with σ=1). Once all the points are transformed, their distributions are summed together and discretized into an image.

The weighting ensures points near the diagonal have a small contribution. This ensures this representation of the diagram is stable.

Once a PersistenceImage is constructed (see below), it can called like a function to transform a diagram to an image.

Infinite intervals in the diagram are ignored.

Constructors

PersistenceImage(ylims, xlims; size=5, kwargs...)

PersistenceImage(diagrams; size=5, kwargs...)

Arguments

ylims, xlims: Limits of the square on which the image is created, both 2-tuples. Note that y comes first as this is the way arrays are indexed.

diagrams: Collection of persistence diagrams. This constructor sets ylims and xlims according to minimum and maximum birth time and persistence time. Sets slope_end to maximum persistence time.

Keyword Arguments

distribution: A function or callable object used to smear each interval in diagram. Has to be callable with two Float64s as input and should return a Float64. Defaults to a normal distribution with sigma equal to 1.

sigma: The width of the gaussian distribution. Only applicable when distribution is unset.

weight: A function or callable object used as the weighting function. Has to be callable with two Float64s as input and should return a Float64. Should equal 0.0 for x=0, but this is not enforced.

slope_end: the $y$ value at which the default weight function stops increasing.

size: integer or tuple of two integers. Determines the size of the array containing the image. Defaults to 5.

Example

julia> diag_1 = PersistenceDiagram([(0, 1), (0, 1.5), (1, 2)]);

julia> diag_2 = PersistenceDiagram([(1, 2), (1, 1.5)]);

julia> image = PersistenceImage([diag_1, diag_2])
5×5 PersistenceImage(
  distribution = PersistenceDiagrams.Binormal(1.0),
  weight = PersistenceDiagrams.DefaultWeightingFunction(1.5),
)

julia> image(diag_1)
5×5 Array{Float64,2}:
 0.266562  0.269891  0.264744  0.251762  0.232227
 0.294472  0.297554  0.291244  0.276314  0.254244
 0.31342   0.316057  0.308664  0.292136  0.268117
 0.32141   0.323446  0.315164  0.297554  0.272373
 0.31758   0.318928  0.310047  0.29199   0.266562

Reference

Adams, H., Emerson, T., Kirby, M., Neville, R., Peterson, C., Shipman, P., ... & Ziegelmeier, L. (2017). Persistence images: A stable vector representation of persistent homology. The Journal of Machine Learning Research, 18(1), 218-252.

PersistenceDiagrams.PersistenceCurve — Type

PersistenceCurve

Persistence curves offer a general way to transform a persistence diagram into a vector of numbers.

This is done by first splitting the time domain into buckets. Then the intervals contained in the bucket are collected and transformed by applying fun to each of them. The result is then summarized with the stat function. If an interval is only parially contained in a bucket, it is counted partially.

Once a PersistenceCurve is constructed (see below), it can be called to convert a persistence diagram to a vector of floats.

Constructors

PersistenceCurve(fun, stat, start, stop; length=10, integrate=true, normalize=false): length buckets with the first strating on t_start and the last ending on t_end.

PersistenceCurve(fun, stat, diagrams; length=10, integreate=true, normalize=false): learn the start and stop parameters from a collection of persistence diagrams.

Keyword arguments

length: the length of the output. Defaults to 10.

fun: the function applied to each interval. Must have the following signature. fun(::AbstractPersistenceInterval, ::PersistenceDiagram, time)::T

stat: the summary function applied the results of fun. Must have the following signature. stat(::Vector{T})::Float64

normalize: if set to true, normalize the result. Does not work for time-dependent funs. Defaults to false. Normalization is performed by dividing all values by stat(fun.(diag)).

integrate: if set to true, the amount of overlap between an interval and a bucket is considered. This prevents missing very small bars, but does not work correctly for curves with time-dependent funs where stat is a selection function (such as landscapes). If set to false, the curve is simply sampled at midpoints of buckets. Defaults to true.

Call

(::PersistenceCurve)(diagram; normalize, integrate)

Transforms a diagram. normalize and integrate override defaults set in constructor.

Example

julia> diagram = PersistenceDiagram([(0, 1), (0.5, 1), (0.5, 0.6), (1, 1.5), (0.5, Inf)]);

julia> curve = BettiCurve(0, 2, length = 4)
PersistenceCurve(always_one, sum, 0.0, 2.0; length=4, normalize=false, integrate=true)

julia> curve(diagram)
4-element Array{Float64,1}:
 1.0
 3.2
 2.0
 1.0

See Also

The following are equivalent to PersistenceCurve with appropriately selected fun and stat arguments.

BettiCurve

Landscape

Silhuette

Life

Midlife

LifeEntropy

MidlifeEntropy

PDThresholding

More options listed in Table 1 on page 9 of reference.

Reference

Chung, Y. M., & Lawson, A. (2019). Persistence curves: A canonical framework for summarizing persistence diagrams. arXiv preprint arXiv:1904.07768.

PersistenceDiagrams.BettiCurve — Function

BettiCurve

Betti curves count the Betti numbers at each time step. Unlike most vectorization methods, they support infinite intervals.

fun(_, _, _) = 1.0
stat = sum

See also

PersistenceCurve

PersistenceDiagrams.Landscape — Function

Landscape(k, args...)

The k-th persistence landscape.

fun((b, d), _, t) = max(min(t - b, d - t), 0)
stat = get(sort(values, rev=true), k, 0.0)

See also

PersistenceCurve

Reference

Bubenik, P. (2015). Statistical topological data analysis using persistence landscapes. The Journal of Machine Learning Research, 16(1), 77-102.

PersistenceDiagrams.Silhuette — Function

Silhuette

The sum of persistence landscapes for all values of k.

fun((b, d), _, t) = max(min(t - b, d - t), 0)
stat = sum

See also

PersistenceCurve Landscape

PersistenceDiagrams.Life — Function

Life

The life curve.

fun((b, d), _, _) = d - b
stat = sum

See also

PersistenceCurve

Reference

Chung, Y. M., & Lawson, A. (2019). Persistence curves: A canonical framework for summarizing persistence diagrams. arXiv preprint arXiv:1904.07768.

PersistenceDiagrams.Midlife — Function

Midlife

The midlife curve.

fun((b, d), _, _) = (b + d) / 2
stat = sum

See also

PersistenceCurve

Reference

Chung, Y. M., & Lawson, A. (2019). Persistence curves: A canonical framework for summarizing persistence diagrams. arXiv preprint arXiv:1904.07768.

PersistenceDiagrams.LifeEntropy — Function

LifeEntropy

The life entropy curve.

fun((b, d), diag, _) = begin
    x = (d - b) / sum(d - b for (b, d) in diag)
    -x * log2(x)
end
stat = sum

See also

PersistenceCurve

Reference

Atienza, N., González-Díaz, R., & Soriano-Trigueros, M. (2018). On the stability of persistent entropy and new summary functions for TDA. arXiv preprint arXiv:1803.08304.

PersistenceDiagrams.MidlifeEntropy — Function

MidlifeEntropy

The midlife entropy curve.

fun((b, d), diag, _) = begin
    x = (b + d) / sum(b + d for (d, b) in diag)
    -x * log2(x)
end
stat = sum

See also

PersistenceCurve

Reference

Chung, Y. M., & Lawson, A. (2019). Persistence curves: A canonical framework for summarizing persistence diagrams. arXiv preprint arXiv:1904.07768.

PersistenceDiagrams.PDThresholding — Function

PDThresholding

The persistence diagram thresholding function.

fun((b, d), _, t) = (d - t) * (t - b)
stat = mean

See also

PersistenceCurve

Reference

Chung, Y. M., & Day, S. (2018). Topological fidelity and image thresholding: A persistent homology approach. Journal of Mathematical Imaging and Vision, 60(7), 1167-1179.