Persistence Intervals and Diagrams

PersistenceDiagrams.PersistenceInterval — Type

PersistenceInterval{T, C}

The type that represents a persistence interval. It behaves exactly like a Tuple{T, Union{T, Infinity}}, but may have a representative cocycle attached to it.

PersistenceDiagrams.birth — Method

birth(interval::PersistenceInterval)

Get the birth time of interval.

PersistenceDiagrams.death — Method

death(interval::PersistenceInterval)

Get the death time of interval.

PersistenceDiagrams.persistence — Method

death(interval::PersistenceInterval)

Get the persistence of interval, which is equal to death - birth.

PersistenceDiagrams.representative — Method

representative(interval::PersistenceInterval)

Get the representative cocycle attached to interval. If representatives were not computed, throw an error.

PersistenceDiagrams.PersistenceDiagram — Type

PersistenceDiagram{P<:PersistenceInterval} <: AbstractVector{P}

Type for representing persistence diagrams. Behaves exactly like an array of PersistenceIntervals, but is aware of its dimension and supports pretty printing and plotting.

PersistenceDiagrams.barcode — Method

barcode(diagram; infinity=nothing)

Plot the barcode plot or AbstractVector of diagrams. The infinity keyword argument determines where the infinity line is placed. If set to nothing the function tries to guess a good infinity poistion.

Distances Between Persistence Diagrams

PersistenceDiagrams.matching — Function

matching(m::Matching; bottleneck=m.bottleneck)

Get the matching of a Matching object represented by a vector of pairs of intervals. If bottleneck is set to true, only return the edges with length equal to the weight of the matching.

matching(::Bottleneck, diag1, diag2)

Find the bottleneck matching between persistence diagrams diag1 and diag2. Infinite intervals are ignored.

diag1 = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
diag2 = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
matching(Bottleneck(), diag1, diag2)

# Example

# output

3-element Matching with weight 2.0:
 [1.0, 2.0) => [1.0, 2.0)
 [3.0, 3.0) => [3.0, 4.0)
 [5.0, 8.0) => [5.0, 10.0)

See also

Bottleneck

distance

matching(::Wasserstein, diag1, diag2)

Find the Wasserstein matching between persistence diagrams diag1 and diag2. Infinite intervals are ignored.

Example

diag1 = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
diag2 = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
matching(Wasserstein(), diag1, diag2)

# output

3-element Matching with weight 3.0:
 [1.0, 2.0) => [1.0, 2.0)
 [3.0, 3.0) => [3.0, 4.0)
 [5.0, 8.0) => [5.0, 10.0)

See also

Wasserstein

distance

PersistenceDiagrams.distance — Function

distance(::Matching)

Get the weight of a Matching object.

distance(::Bottleneck, diag1, diag2)

Compute the bottleneck distance between persistence diagrams diag1 and diag2. Infinite intervals are ignored.

Example

diag1 = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
diag2 = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
distance(Bottleneck(), diag1, diag2)

# output

2.0

See also

Bottleneck

matching

distance(::Wasserstein, diag1, diag2)

Compute the Wasserstein distance between persistence diagrams diag1 and diag2. Infinite intervals are ignored.

Example

diag1 = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
diag2 = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
distance(Wasserstein(), diag1, diag2)

# output

3.0

See also

Wasserstein

matching

PersistenceDiagrams.Bottleneck — Type

Bottleneck

Use this object to find the bottleneck distance or matching between persistence diagrams. The distance value is equal to

\[W_\infty(X, Y) = \inf_{\eta:X\rightarrow Y} \sup_{x\in X} ||x-\eta(x)||_\infty,\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the bottleneck distance requires $\mathcal{O}(n^2)$ space!

Methods

matching(::Bottleneck, ::Any, ::Any): construct a bottleneck Matching.

distance(::Bottleneck, ::Any, ::Any): find the bottleneck distance.

PersistenceDiagrams.Wasserstein — Type

Wasserstein(q=1)

Use this object to find the Wasserstein distance or matching between persistence diagrams. The distance value is equal to

\[W_q(X,Y)=\left[\inf_{\eta:X\rightarrow Y}\sum_{x\in X}||x-\eta(x)||_\infty^q\right],\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the Wasserstein distance requires $\mathcal{O}(n^2)$ space!

Methods

matching(::Wasserstein, ::Any, ::Any): construct a Wasserstein Matching.

distance(::Wasserstein, ::Any, ::Any): find the Wasserstein distance.

PersistenceDiagrams.Matching — Type

Matching

A matching between two persistence diagrams.

Methods

distance(::Matching)

matching(::Matching)