Persistence Intervals and Diagrams

PersistenceDiagrams.PersistenceInterval — Type

PersistenceInterval{T<:AbstractFloat, C}

The type that represents a persistence interval. It behaves exactly like a Tuple{Float64, Float64}.

PersistenceDiagrams.birth — Function

birth(interval)

Get the birth time of interval.

PersistenceDiagrams.death — Function

death(interval)

Get the death time of interval.

PersistenceDiagrams.persistence — Function

persistence(interval)

Get the persistence of interval, which is equal to death - birth.

PersistenceDiagrams.RepresentativeInterval — Type

RepresentativeInterval{P<:AbstractInterval, B, D, R} <: AbstractInterval

A persistence interval with a representative (co)cycles and critical simplices attached.

PersistenceDiagrams.representative — Function

representative(interval::RepresentativeInterval)

Get the representative (co)cycle attached to interval.

PersistenceDiagrams.birth_simplex — Function

birth_simplex(interval::RepresentativeInterval)

Get the critical birth simplex of interval.

PersistenceDiagrams.death_simplex — Function

death_simplex(interval::RepresentativeInterval)

Get the critical death simplex of interval.

PersistenceDiagrams.PersistenceDiagram — Type

PersistenceDiagram{P<:AbstractInterval} <: AbstractVector{P}

Type for representing persistence diagrams. Behaves exactly like an array of AbstractIntervals, but is aware of its dimension and supports pretty printing and plotting.

PersistenceDiagrams.barcode — Method

barcode(diagram)

Plot the barcode plot of persistence diagram or multiple diagrams diagrams. The infinity keyword argument determines where the infinity line is placed. If unset, the function tries to use threshold(diagram) or guess a good position to place the line at.

Distances Between Persistence Diagrams

PersistenceDiagrams.Bottleneck — Type

Bottleneck

Use this object to find the bottleneck distance or matching between persistence diagrams. The distance value is equal to

\[W_\infty(X, Y) = \inf_{\eta:X\rightarrow Y} \sup_{x\in X} ||x-\eta(x)||_\infty,\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the bottleneck distance requires $\mathcal{O}(n^2)$ space. Be careful when computing distances between very large diagrams!

Usage

Bottleneck()(left, right[; matching=false]): find the bottleneck matching (if matching=true) or distance (if matching=false) between persistence diagrams left and right

Example

left = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
right = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
Bottleneck()(left, right)

# output

2.0

PersistenceDiagrams.Wasserstein — Type

Wasserstein(q=1)

Use this object to find the Wasserstein distance or matching between persistence diagrams. The distance value is equal to

\[W_q(X,Y)=\left[\inf_{\eta:X\rightarrow Y}\sum_{x\in X}||x-\eta(x)||_\infty^q\right],\]

where $X$ and $Y$ are the persistence diagrams and $\eta$ is a perfect matching between the intervals. Note the $X$ and $Y$ don't need to have the same number of points, as the diagonal points are considered in the matching as well.

Warning

Computing the Wasserstein distance requires $\mathcal{O}(n^2)$ space. Be careful when computing distances between very large diagrams!

Usage

Wasserstein(q=1)(left, right[; matching=false]): find the Wasserstein matching (if matching=true) or distance (if matching=false) between persistence diagrams left and right.

Example

left = PersistenceDiagram(0, [(1.0, 2.0), (5.0, 8.0)])
right = PersistenceDiagram(0, [(1.0, 2.0), (3.0, 4.0), (5.0, 10.0)])
Wasserstein()(left, right)

# output

3.0

PersistenceDiagrams.matching — Function

matching(::MatchingDistance, left, right)
matching(::Matching)

Get the matching between persistence diagrams left and right.

See also

matching

Bottleneck

Wasserstein

PersistenceDiagrams.weight — Function

weight(::MatchingDistance, left, right)
weight(::Matching)

Get the weight of the matching between persistence diagrams left and right.

See also

matching

Bottleneck

Wasserstein

Persistence Diagram Vectorization Methods

PersistenceDiagrams.BettiCurve — Type

BettiCurve

A betti curve is a simple way of transforming a persistence diagram to a vector of real numbers. A BettiCurve bc splits the barcode of a diagram into length(bc) buckets, counting the number of bars in each bucket. If only a part of a bar is contained in a bucket, it is only partially counted.

Once a BettiCurve is constructed (see below), it can be called to convert a persistence diagram to a vector of floats.

Unlike most diagram vectorization methods, BettiCurve can handle infinite intervals.

Constructors

BettiCurve(t_start, t_end; length=10): length buckets with the first strating on t_start and the last ending on t_end.

BettiCurve(diagrams; length=10): learn the t_start and t_end parameters from a collection of persistence diagrams.

BettiCurve(buckets): manually select the buckets. buckets should be an AbstractVector of length one higher than the desired output size.

Example

diagram = PersistenceDiagram(0, [(0, 1), (0.5, 1), (0.5, 0.6), (1, 1.5), (0.5, Inf)])
curve = BettiCurve(0, 2, length = 4)
curve(diagram)

# output

4-element Array{Float64,1}:
 1.0
 3.2
 2.0
 1.0

PersistenceDiagrams.PersistenceImage — Type

PersistenceImage

PersistenceImage provides a vectorization method for persistence diagrams. Each point in the diagram is first transformed into birth, persistence coordinates. Then, it is weighted by a weighting function and widened by a distribution (default: gaussian with σ=1). Once all the points are transformed, their distributions are summed together and discretized into an image.

The weighting ensures points near the diagonal have a small contribution. This ensures this representation of the diagram is stable.

Once a PersistenceImage is constructed (see below), it can called like a function to transform a diagram to an image.

Infinite intervals in the diagram are ignored.

Constructors

PersistenceImage(ylims, xlims; size=5, kwargs...)

PersistenceImage(diagrams; size=5, kwargs...)

Arguments

ylims, xlims: Limits of the square on which the image is created, both 2-tuples. Note that y comes first as this is the way arrays are indexed.

diagrams: Collection of persistence diagrams. This constructor sets ylims and xlims according to minimum and maximum birth time and persistence time. Sets slope_end to maximum persistence time.

Keyword Arguments

distribution: A function or callable object used to smear each interval in diagram. Has to be callable with two Float64s as input and should return a Float64. Defaults to a normal distribution with sigma equal to 1.

sigma: The width of the gaussian distribution. Only applicable when distribution is unset.

weight: A function or callable object used as the weighting function. Has to be callable with two Float64s as input and should return a Float64. Should equal 0.0 for x=0, but this is not enforced.

slope_end: the $y$ value at which the default weight function stops increasing.

size: integer or tuple of two integers. Determines the size of the array containing the image. Defaults to 5.

Example

diag_1 = PersistenceDiagram(0, [(0, 1), (0, 1.5), (1, 2)])
diag_2 = PersistenceDiagram(0, [(1, 2), (1, 1.5)])
image = PersistenceImage([diag_1, diag_2])

# output

5×5 PersistenceImage(
  distribution = PersistenceDiagrams.Binormal(1.0),
  weight = PersistenceDiagrams.DefaultWeightingFunction(1.5)
)

image(diag_1)

# output

5×5 Array{Float64,2}:
 0.266562  0.269891  0.264744  0.251762  0.232227
 0.294472  0.297554  0.291244  0.276314  0.254244
 0.31342   0.316057  0.308664  0.292136  0.268117
 0.32141   0.323446  0.315164  0.297554  0.272373
 0.31758   0.318928  0.310047  0.29199   0.266562